pyproject.toml
[project]
name = "delexplore"
version = "0.1.0"
description = "Statistical analysis, hit ranking, and chemical space exploration of DEL screening data"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["DEL", "DNA-encoded library", "cheminformatics", "drug-discovery"]
dependencies = [
    "click>=8.0",
    "polars>=0.20",
    "pyarrow>=14.0",
    "scipy>=1.11",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
# Chemistry: SMILES handling, fingerprints, property calculation, scaffold analysis
chem = [
    "rdkit>=2023.9",
]
# Chemical-space exploration: UMAP, clustering, interactive plots
explore = [
    "umap-learn>=0.5",
    "hdbscan>=0.8",
    "plotly>=5.18",
    "matplotlib>=3.8",
]
# Statistical modelling and ML enrichment
ml = [
    "pydeseq2>=0.4",
    "scikit-learn>=1.3",
]
# Development tooling
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.0",
]
# Install everything
all = ["delexplore[chem,explore,ml,dev]"]

[project.scripts]
delexplore = "delexplore.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/delexplore"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
